+++ apps/ngrid-demo-app/content/demos/infinite-scroll-performance-demo/infinite-scroll-performance-demo.component.ts
@@ -128,6 +128,7 @@ export class InfiniteScrollPerformanceDemoExample {
         initialVirtualSize: 100,
       })
       .withCacheOptions('sequenceBlocks')
+      .setCustomTriggers('filter')
       .onTrigger( event => {
         if (event.isInitial) {
           return this.datasource.getCountries()