+++ apps/ngrid-demo-app/content/features/grid/infinite-scroll/custom-triggers.component.ts
@@ -1,4 +1,3 @@
-import { Observable } from 'rxjs';
 import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
 import { columnFactory } from '@pebula/ngrid';
 import { createInfiniteScrollDS } from '@pebula/ngrid/infinite-scroll';
@@ -7,20 +6,19 @@ import { Person, NgridDemoRestApiClient } from '@pebula/apps/shared-data';
 import { Example } from '@pebula/apps/shared';
 
 @Component({
-  selector: 'pbl-index-based-paging-example',
-  templateUrl: './index-based-paging.component.html',
-  styleUrls: ['./index-based-paging.component.scss'],
+  selector: 'pbl-custom-triggers-example',
+  templateUrl: './custom-triggers.component.html',
+  styleUrls: ['./custom-triggers.component.scss'],
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-@Example('pbl-index-based-paging-example', { title: 'Index Based Paging' })
-export class IndexBasedPagingExample {
-
+@Example('pbl-custom-triggers-example', { title: 'Custom Triggers' })
+export class CustomTriggersExample {
   columns = columnFactory()
   .table(
     { prop: 'id', width: '100px', pIndex: true },
-    { prop: 'name', width: '100px', editable: true },
-    { prop: 'gender', width: '50px' },
+    { prop: 'name', width: '100px', sort: true },
+    { prop: 'gender', width: '50px', sort: true },
     { prop: 'birthdate', type: 'date', width: '25%' },
   )
   .build();
@@ -31,6 +29,7 @@ ds = createInfiniteScrollDS<Person>()
     initialVirtualSize: 100,
   })
   .withCacheOptions('fragmentedBlocks', { strictPaging: false })
+  .setCustomTriggers('filter', 'sort') // pagination is simply ignored in this case
   .onTrigger(event => {
     if (event.isInitial) {
       this.ds.setCacheSize(200 * 4);
@@ -41,18 +40,21 @@ ds = createInfiniteScrollDS<Person>()
         return resp.items;
       });
     } else {
-      // At this point "event.fromRow" will never be >= to event.totalLength
-      // We defined the total length in the first trigger, so the datasource will make sure to avoid requests
-      // for data out of range.
-
-      console.log(event.fromRow, event.toRow);
-
-      return this.client.getPeople({
-        pagination: { skip: event.fromRow, limit: event.toRow - event.fromRow + 1},
-      })
-      .then( resp => {
-        return resp.items;
-      });
+      switch (event.eventSource) {
+        case 'infiniteScroll':
+          return this.client.getPeople({
+            pagination: { skip: event.fromRow, limit: event.toRow - event.fromRow + 1},
+          }).then( resp => resp.items );
+        case 'customTrigger':
+          if (event.sort.changed) {
+            // handle sort, go to server and ask to sort
+          } else if (event.filter.changed) {
+            // handle filter, go to server and ask to filter
+          }
+          break;
+        default:
+          throw new Error('This should NEVER EVENT happen...');
+      }
     }
   })
   .create();